+++ packages/sqlite-worker/src/create/Sets.ts
@@ -1,5 +1,6 @@
 import type { MigrationQueryBus } from '..';
 import type Sqlite3 from '../Sqlite3'
+import { isNewVersion } from './Version';
 
 const createSetsTableSql = `CREATE TABLE IF NOT EXISTS sets (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
@@ -18,26 +19,24 @@ export default function createSetTable(db: Sqlite3) {
 }
 
 export function migrate(db: MigrationQueryBus, v: Versions) {
-  switch (v) {
-    case '1.1.0': {
-      const queryList = db.get('1.1.0') || []
-      queryList.push({
-        sql: 'alter table sets rename to sets_old',
-        args: []
-      })
-      queryList.push({
-        sql: createSetsTableSql,
-        args: []
-      })
-      queryList.push({
-        sql: 'insert into sets (exerciseId, repeat, isDone, weightUnit, weight, duration) select exerciseId, repeat, isDone, weightUnit, weight, duration from sets_old',
-        args: []
-      })
-      queryList.push({
-        sql: 'drop table sets_old',
-        args: []
-      })
-      db.set('1.1.0', queryList)
-    } break
+  if (isNewVersion(v, '1.3.0')) {
+    const queryList = db.get('1.3.0') || []
+    queryList.push({
+      sql: 'alter table sets rename to sets_old',
+      args: []
+    })
+    queryList.push({
+      sql: createSetsTableSql,
+      args: []
+    })
+    queryList.push({
+      sql: 'insert into sets (exerciseId, repeat, isDone, weightUnit, weight, duration) select exerciseId, repeat, isDone, weightUnit, weight, duration from sets_old',
+      args: []
+    })
+    queryList.push({
+      sql: 'drop table sets_old',
+      args: []
+    })
+    db.set('1.3.0', queryList)
   }
 }
